# backend/docs/swagger.yaml
definitions:
  Controllers.Resp:
    properties:
      data:
        type: object
      error:
        type: string
      status:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: withrjp@gmail.com
    name: Ackerman
    url: www.withrjp.xyz
  description: Ackerman Blog Api Documents
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ackerman Blog
  version: "1.0"
paths:
...
  /api/users/{id}:
    get:
      consumes:
      - application/json
      description: 根据用户id查询用户信息
      operationId: get-string-by-int
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Controllers.Resp'
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 用户信息查询
      tags:
      - user
swagger: "2.0"